Обозначает файловую структуру проекта: добавляет файлы для будущих js-модулей
