+++ index.html
@@ -230,6 +230,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <script src="./js/main.js" defer></script>
+  <script type="module" src="./js/main.js" defer></script>
 </body>
 </html>
